Separate queries and commands into folders
