+++ Application/Items/Queries/Details.cs
@@ -9,25 +9,22 @@ using System.Threading;
 using System.Threading.Tasks;
 using Application.Exceptions;
 
-namespace Application.Items
+namespace Application.Items.Queries
 {
-    public class Delete
+    public class Details
     {
-        public class Command : IRequest
+        public class Query : IRequest<Item>
         {
             public Guid Id { get; set; }
-            public string User { get; set; }
         }
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Query, Item>
         {
             private readonly ItemDbContext _context;
-
             public Handler(ItemDbContext context)
             {
                 _context = context;
             }
-
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Item> Handle(Query request, CancellationToken cancellationToken)
             {
                 var item = await _context.Items.FindAsync(request.Id);
 
@@ -36,16 +33,7 @@ namespace Application.Items
                     throw new NotFoundException(nameof(Item), request.Id);
                 }
 
-                if (item.Uploader != request.User)
-                {
-                    throw new ConflictException($"Item {request.Id} does not belong to the client");
-                }
-                
-                _context.Remove(item);
-                
-                await _context.SaveChangesAsync();
-                
-                return Unit.Value;
+                return item;
             }
         }
     }
